server/requirements.txt
docxtpl==0.20.0
docx2pdf==0.1.8
google-api-python-client==2.175.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
flask==3.1.1
flask-cors==6.0.1